requirements.txt
selenium
webdriver-manager
ipykernel
openpyxl
pandas

aniso8601
click
Flask
Flask-Cors
Flask-RESTful
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
Flask-Wtf
Wtforms
Email-Validator
itsdangerous
Jinja2
MarkupSafe
pytz
six
SQLAlchemy
Werkzeug
mysql-connector-python